model-api/pyproject.toml
[tool.poetry]
name = "model-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-restx = "^1.3.0"
flask = "^3.0.2"
torch = "^2.2.1"
torchvision = "^0.17.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
